+++ NEWS.md
@@ -1,3 +1,20 @@
+# CHANGES IN typr VERSION 0.2
+
+## NEW FEATURES
+
+- added: argument `return_invis` in `setTyped()`:
+  allows the fine control of the actual return values
+
+## BUG FIXES
+
+## MAJOR CHANGES
+
+## MINOR CHANGES
+
+## MISC
+
+-----
+
 # CHANGES IN nestr VERSION 0.1
 
 ## NEW FEATURES